+++ pages/api/examples/curriculum-vitae.ts
@@ -29,8 +29,27 @@ export function response() {
     ],
     experiences: [
       {
-        id: "5bf6539c-2872-477b-93a5-bf6ed8754a0g",
+        id: "1bc0fd31-9950-4441-b7f1-6975aa53b574",
         priority: true,
+        role: "Software Engineer",
+        company: "Lloyd's Banking Group",
+        tenure: "November '22 - Present",
+        text: [
+          {
+            id: "3afacc94-e5ca-444f-91c5-b1d5d8b79633",
+            value:
+              "At Lloyd's Banking Group, I play a key role in delivering best-in-class mortgage journeys to customers, utilizing React and TypeScript to build modern, digital experiences. As the lead engineer, I managed a team responsible for migrating from legacy applications to new digital solutions, ensuring that we achieved optimal outcomes for customers. A notable achievement was leading the replatforming of a legacy application, completing the project ahead of schedule within a very tight timeframe."
+          },
+          {
+            id: "2faa1093-e469-4041-b1a3-29c359498e10",
+            value:
+              "In recognition of my contributions, I was honored with the platform-wide 'Trust' award at the end of 2024, nominated by my peers for my dedication to stepping up and supporting both my leaders and my team. I served as a key enabler, expediting work across the platform and driving progress across various environments. In addition to my technical leadership, I also mentor an industry placement engineer, helping to nurture and grow the group's talent footprint while fostering a collaborative and supportive team environment."
+          }
+        ]
+      },
+      {
+        id: "5bf6539c-2872-477b-93a5-bf6ed8754a0g",
+        priority: false,
         role: "Software Development Engineer",
         company: "EMIS Health",
         tenure: "November '17 - November '22",
@@ -38,32 +57,22 @@ export function response() {
           {
             id: "0b3db32f-a3f3-4361-8538-cbf7c2c2eb2a",
             value:
-              "I have enjoyed working with a wide ranging, ever changing tech stack during my time at the company which began in C#, involving the refactoring of a legacy product alongside resolving priority SLA items."
+              "At EMIS Health, I worked with a diverse and ever-evolving tech stack, beginning with .NET development on their flagship product. I played a key role in ensuring high-priority software licensing agreements were met ahead of deadlines."
           },
           {
             id: "69c06caa-635e-421e-94b8-3953afcf3c9b",
             value:
-              "I was a member of the first set of engineers to begin coding the next generation of product for the business which allowed a great deal of flexibility and freedom as we identified new tools and services fit for use. Including as a learning exercise the creation of a Twitter bot (@motivationalboz) using Node.js and AWS Lambda to post a tweet on a CRON schedule."
+              "I was part of the first engineering team tasked with building the next-generation product, which provided the opportunity to transition into a variety of technologies, including React, TypeScript, and Node.js. This shift gave me the flexibility to explore and implement innovative tools and services, a key achievemnt was the creation of an ‘X’ bot (@motivationalboz) using Node.js and AWS Lambda, automating tweets on a CRON schedule as part of a learning exercise."
           },
           {
             id: "26d2935f-5d9b-441f-ac16-aef47c4daedc",
             value:
-              "After this period of elaboration I began to work exclusively in this space with Golang and the varying services AWS provide, inc. API Gateway, CloudFormation, CodePipeline and Lambda. Writing one of the first CloudFormation templates which deployed a CI/CD pipeline in the department."
+              "As my role evolved, I expanded my expertise further into front-end technologies like React and TypeScript, adding tools including Terraform and Docker for Infrastructure as Code (IaC), repository management, and containerizing applications. This experience was instrumental in the development of industry-leading solutions."
           },
           {
             id: "44446faf-11a5-4c2f-a83a-b40442fff3f7",
             value:
-              "This has organically led to the consumption of API’s and a more front-end focus with the use of React and TypeScript, and allowed for more hands-on experience with the likes of Terraform and Docker through repository management and containerised applications."
-          },
-          {
-            id: "06995db5-aed9-49bf-a831-a4ee6a49ca17",
-            value:
-              "I began to contribute to our new and growing community initiatives before identifying an opportunity and leading on my own with the support of my peers, whilst actively looking for areas of improvement not only within the team but within the department by identifying new tools and best practices."
-          },
-          {
-            id: "145305f3-717b-4fc3-b147-36c3444aa906",
-            value:
-              "I naturally undertook the mentoring of an apprentice developer and test engineer within my team, which led to a company-wide award in the inaugural EMIS Quarterly Heroes Q1 2021 for living the Supportive value of the business."
+              "I actively contributed to the company’s community initiatives, identifying opportunities for improvement across both my team and department. This led me to mentor both an apprentice developer and a test engineer, fostering knowledge sharing and growth within the team. My leadership and proactive approach were recognized with a company-wide award at the inaugural EMIS Quarterly Heroes in Q1 2021, where I was honored for embodying the 'Supportive' value of the business."
           }
         ]
       },
@@ -77,7 +86,7 @@ export function response() {
           {
             id: "fb72a0cd-228f-4c5a-8053-72f51f3107a8",
             value:
-              "I joined the award winning training programme and undertook weekly modules such as: SQL, Web Development and Professional Skills. Before a more concentrated .NET period which focused on the SOLID principles, Object-oriented programming and TDD. Culminating in a presentation to my peers, and wider department involving the eCommerce project created during this time."
+              "I joined FDM Group’s award-winning training program, where I gained hands-on experience through weekly modules covering a variety of key topics such as SQL, Web Development, and Professional Skills. My training culminated in a focused period on .NET development, where I deepened my understanding of core principles such as SOLID, Object-Oriented Programming (OOP), and Test-Driven Development (TDD). This practical learning was applied to an eCommerce project, which I presented to both my peers and wider department, showcasing my ability to apply industry-standard techniques to real-world development challenges."
           }
         ]
       },
@@ -91,7 +100,7 @@ export function response() {
           {
             id: "d2b8ad77-e28c-422c-9819-22ba2ad0e22a",
             value:
-              "Responsible for delivering a first class customer service in any capacity around the store. My role was to in the first instance operate the checkouts, whilst during busier periods I would become the focal point of the checkouts department. Ensuring all colleagues have everything they need to hand, answering any customer queries a colleague could not."
+              "In this role, I was responsible for delivering a first-class customer experience both at the tills and on self-service checkouts. I consistently provided excellent service, assisting customers efficiently while ensuring their needs were met with a friendly and professional approach. As the leader within the checkout department, I was tasked with managing and supporting a team of colleagues. This included ensuring they understood their responsibilities and were well-equipped to contribute to the department's smooth operations. My leadership helped ensure that customers had a seamless journey through the department, contributing to the overall positive shopping experience."
           }
         ]
       }
@@ -124,12 +133,12 @@ export function response() {
         {
           id: "11dc047d-9f34-4557-938a-01294d74e81a",
           value:
-            "An experienced software engineer with a passion for developing web applications. Well-versed in a variety of technologies to create applications that are reliable and user-friendly."
+            "An experienced software engineer in delivering reliable, secure, innovative, customer-focused solutions."
         },
         {
           id: "113945ae-5d6c-4233-9957-25d66a5ebdab",
           value:
-            "A budding leader with a proven ability to mentor and motivate colleagues, who can communicate confidently, problem-solve effectively and use initiative to deliver applications tailored to organizational needs using best practices and the best tool for the job."
+            "A proactive leader and mentor with proven experiences in improving team processes and delivery, earning company-wide awards across roles for my supportive contributions."
         }
       ]
     }